==> notebooks/02_requests_apis.ipynb <==
- Sacar datos luz REAPI por mes (2019-2022): https://apidatos.ree.es/es/datos/mercados/componentes-precio?start_date=2019-01-01T00:00&end_date=2019-01-31T23:59&time_trunc=month&geo_limit=ccaa
    - El precio en españa es el mismo para todos
- Sacar datos gas: https://data.cnmc.es/energia/mercado-mayorista-del-gas-natural/conjuntos-de-datos/evolucion-de-los-precios-por-productos

# Importamos las librerías que necesitamos

# Librerías para el tratamiento y manipulación de datos
# -----------------------------------------------------------------------
import pandas as pd             # Manipulación de datos en estructuras DataFrame para análisis y transformación
import numpy as np              # Realización de cálculos numéricos y manipulación de arrays multidimensionales
import datetime                 # Manejo y procesamiento de fechas y horas, útil para organizar datos temporalmente

# Manejo del tiempo y generación de pausas
# -----------------------------------------------------------------------
from time import sleep          # Pausa la ejecución del código durante un tiempo determinado, útil para evitar bloqueos en scraping
import random                   # Generación de valores aleatorios, ideal para espaciar solicitudes en procesos de scraping

# Para importar módulos desde rutas recursivas
# -----------------------------------------------------------------------
import sys
sys.path.append("../")          # Añade el directorio padre al PATH para permitir la importación de módulos personalizados
from src.SupportApi import (    # Importa funciones personalizadas de soporte
    api_gas,                    # Función para hacer llamadas a la API de gas
    api_luz                     # Función para hacer llamadas a la API de electricidad
)

# Control de advertencias
# -----------------------------------------------------------------------
import warnings
warnings.filterwarnings('ignore')  # Suprime advertencias para mantener la salida de consola más limpia

# Para realizar solicitudes HTTP
import requests                  # Realiza solicitudes HTTP a APIs y obtiene datos en formato JSON


# Obtener Precios gas

Empezamos por obtener todos los precios del gas de 2019 a 2022

la función de abajo nos permite obtener los datos de 2019 a 2022. o en un rango de 2015 a 2023 que son los datos que se tienen hasta la fecha

df_gas = api_gas("2018-12-30","2022-12-01")
df_gas

Ahora vamos a dejar solo el producto de mibgas que nos interesa actualmente los productos que hay son:

df_gas["producto_de_mibgas"].unique()

Vamos entonces a filtrar por el producto mensual M + 1, que nos va a proveer del precio del gas en el mes siguiente (Month + 1)

cond = df_gas["producto_de_mibgas"] == "Producto mensual M+1"
df_gas = df_gas[cond]

Lo guardamos, con objetivo de guardar nuestro progreso actual

df_gas.to_csv("../datos/01_PreciosGas/01_PrecioGasMensualSinLimpiar2019-2022.csv")

## Limpieza Datos Gas
Dado que hay múltiples precios en el mismo mes, nos vamos a quedar con el último precio registrado de cada mes

Primero ordenamos por fecha para asegurarnos de que están ordenados

df_gas = df_gas.sort_values(by="fecha_de_negociacion")

Ahora, creamos una nueva columna que nos deja solo el mes y el año

df_gas["anio_mes"] = df_gas["fecha_de_negociacion"].dt.to_period("M")
df_gas.sample()

Agrupamos por el último precio de cada mes y reseteamos el index, también generamos un index compatible para 

df_gas = df_gas.groupby("anio_mes").tail(1).reset_index(drop=True)
df_gas.index = df_gas.index + 1
df_gas.reset_index(inplace=True)
df_gas.sample()

Guardamos nuestro progreso y continuamos

df_gas.to_csv("../datos/01_PreciosGas/02_PrecioGasPorMesSinLimpiar.csv")

Ahora que ya tenemos los precios, por simplicidad, vamos a marcar correctamente a que mes y año pertenecen para poder limpiar más los datos, hacemos un offset de un mes

df_gas["fecha_de_negociacion"] = df_gas["fecha_de_negociacion"] + pd.DateOffset(months=1)

### Eliminamos las columnas innecesarias y generamos correctamente la columna aniomes

df_gas.drop(columns=["_id","producto_de_mibgas","anio_mes"],inplace=True)
df_gas.head()

df_gas["fecha_de_negociacion"] = pd.to_datetime(df_gas["fecha_de_negociacion"]).dt.to_period("M").dt.to_timestamp()

df_gas = df_gas.rename(columns={"fecha_de_negociacion": "fecha",
                                "precios_de_los_productos_mibgas":"precio_gas"})
df_gas = df_gas.assign(fecha=df_gas.pop("fecha"))
df_gas.head()

df_gas.info()

Guardamos como producto Final y pasamos a la siguiente API

df_gas.to_csv("../datos/01_PreciosGas/03_TablaPreciosGasFinal.csv")

# REData API (Red Eléctrica de España)

Vamos a obtener el precio mensual de la electricidad en España de 2019 a 2022 para ver la evolución
para ello haremos peticiones a esta url:
- https://apidatos.ree.es/es/datos/mercados/componentes-precio?start_date=2019-01-01T00:00&end_date=2019-01-31T23:59&time_trunc=month&geo_limit=ccaa

Donde vamos a modificar start_date y end_date en varias ocasiones para poder obtenerlo, ya que por limitación, la API solo nos da de año en año
Con la siguiente función unimos cada request al DataFrame, donde solo nos interesa el precio y el año y mes

years = [2019,2020,2021,2022]
df_luz = api_luz(years,"../datos/02_PreciosLuz")

df_luz.head()

df_luz.info()

### Pasar a formato Datetime

df_luz["fecha"] = df_luz["fecha"].str.split("T").str[0]
df_luz["fecha"] = pd.to_datetime(df_luz["fecha"],format="%Y-%m-%d",utc=False)
df_luz.head()

df_luz.info()

### Lo Guardamos y continuamos

df_luz.to_csv("../datos/02_PreciosLuz/02_PreciosLuzFinal.csv")
