==> notebooks/03_preparacion_preDB.ipynb <==
# Fase PreDB

El objetivo es preparar las tablas para nuestra Base de Datos
Vamos a tener las siguientes tablas y en el siguiente orden:
- fechas
- combustibles
- categorías
- unidades
- historico_precios

# Librerías para el tratamiento y manipulación de datos
# -----------------------------------------------------------------------
import pandas as pd             # Manipulación de datos en estructuras DataFrame para análisis y transformación
import numpy as np              # Realización de cálculos numéricos y manipulación de arrays multidimensionales
import datetime                 # Manejo y procesamiento de fechas y horas, útil para organizar datos temporalmente

# Librerías del sistema y utilidades
# -----------------------------------------------------------------------
import os                       # Interactúa con el sistema operativo para gestionar rutas y archivos
import sys                      # Modifica las rutas de búsqueda de módulos y permite gestionar la ejecución del programa
sys.path.append("../")          # Añade el directorio padre al PATH para importar módulos personalizados

# Importación de funciones personalizadas para preparar datos antes de la carga en la base de datos
from src.SupportPreDB import (
    crear_dictio,               # Crea un diccionario para mapear valores en un DataFrame
    df_a_db                     # Mapea valores en un DataFrame usando un diccionario, preparando los datos para la base de datos
)

df_gas = pd.read_csv("../datos/01_PreciosGas/03_TablaPreciosGasFinal.csv",index_col="Unnamed: 0")
df_luz = pd.read_csv("../datos/02_PreciosLuz/02_PreciosLuzFinal.csv",index_col="Unnamed: 0")
df_gasolina = pd.read_csv("../datos/03_PreciosCombustible/01_TablaPreciosCombustible.csv",index_col="Unnamed: 0")
df_agua = pd.read_csv("../datos/04_PreciosAgua/01_PreciosAguaFinal.csv",index_col="Unnamed: 0")

# Tabla Fechas
Para esta tabla cogemos los valores únicos de las fechas de cualquiera de nuestras tablas

df_fechas = pd.DataFrame(df_gas["fecha"].unique())
df_fechas.index = df_fechas.index + 1
df_fechas.reset_index(inplace=True)
df_fechas = df_fechas.rename(columns={0:"fecha"})
df_fechas.sample()

df_fechas.to_csv("../datos/05_TablasPreDB/01_tabla_fechas.csv")

# Tabla combustibles
Para esta tabla sacamos los valores únicos del tipo combustible de la tabla "Combustibles"

df_combustibles = pd.DataFrame(df_gasolina["combustible"].unique())
df_combustibles = df_combustibles.rename(columns={0:"combustible"})
df_combustibles = pd.concat([pd.DataFrame({"index": [0], "combustible": ["N/A"]}), df_combustibles], ignore_index=True)
df_combustibles = df_combustibles.drop(columns="index")
df_combustibles.index = df_combustibles.index + 1
df_combustibles.reset_index(inplace=True)
df_combustibles

df_combustibles.to_csv("../datos/05_TablasPreDB/02_tabla_combustibles.csv")

# Tabla categorías
Categorías de los valores que estamos comprobando:
- luz
- gas
- combustible
- agua

categorias = ["luz","gas","combustible","agua"]
df_categorias = pd.DataFrame(categorias)
df_categorias.index = df_categorias.index + 1
df_categorias.reset_index(inplace=True)
df_categorias = df_categorias.rename(columns={0:"categoria"})
df_categorias

df_categorias.to_csv("../datos/05_TablasPreDB/03_tabla_categorias.csv")

# Tabla unidades
cada Precio tiene su unidad que son:
- €/L para la gasolina
- €/m3 para el agua
- €/MWh para la luz y el Gas

unidades = ["€/L","€/m3","€/MWh"]
df_unidades = pd.DataFrame(unidades)
df_unidades.index = df_unidades.index + 1
df_unidades.reset_index(inplace=True)
df_unidades = df_unidades.rename(columns={0:"unidad"})
df_unidades

df_unidades.to_csv("../datos/05_TablasPreDB/04_tabla_unidades.csv")

# Tabla Histórico_precios
Para esta tabla haremos un concat de todas las tablas que tenemos, primero las prepararemos para unirlas y después reemplazaremos los valores por ids donde sea necesario

La tabla va a quedar con estas columnas (en todas):
- categoria : Si es luz, gas, combustible o agua
- combustible: en caso de ser combustible si es 95,98, diesel o diesel premium
- precio: el coste en €
- unidad: la unidad del coste en € (si es por litro, metro cúbico o Megavatio hora)
- fecha: La fecha (en este caso el mes)

### Tabla Histórico_luz


df_luz.sample()

df_luz = df_luz.drop(columns="index")
df_luz["categoria"] = "luz"
df_luz["combustible"] = "N/A"
df_luz["unidad"] = "€/MWh"
df_luz = df_luz[["categoria","combustible","precio","unidad","fecha"]]
df_luz.sample()

### Tabla Histórico_gas

df_gas.sample()

df_gas = df_gas.drop(columns="index")
df_gas["categoria"] = "gas"
df_gas["combustible"] = "N/A"
df_gas["unidad"] = "€/MWh"
df_gas = df_gas.rename(columns={"precio_gas":"precio"})
df_gas = df_gas[["categoria","combustible","precio","unidad","fecha"]]
df_gas.sample()

### Tabla Histórico_combustible

df_gasolina.sample()

df_gasolina["categoria"] = "combustible"
df_gasolina["unidad"] = "€/L"
df_gasolina = df_gasolina.rename(columns={"Fecha":"fecha",
                                          "Precio":"precio"})
df_gasolina = df_gasolina[["categoria","combustible","precio","unidad","fecha"]]
df_gasolina.sample()

### Tabla Histórico_agua

df_agua.sample()
df_agua["categoria"] = "agua"
df_agua["combustible"] = "N/A"
df_agua["unidad"] = "€/m3"
df_agua = df_agua.rename(columns={"consumo (€/m3)": "precio"})
df_agua = df_agua[["categoria","combustible","precio","unidad","fecha"]]
df_agua.sample()

# Unir las tablas en una

df_historico = pd.concat([df_luz,df_gas,df_gasolina,df_agua],ignore_index=True)

df_historico.sample()

df_historico.info()

### Ordenar por fecha y añadirle un index (su id de la DB)

df_historico = df_historico.sort_values(by="fecha")
df_historico.reset_index(drop=True,inplace=True)
df_historico.index = df_historico.index + 1
df_historico.reset_index(inplace=True)
df_historico.head()

### Reemplazar categoria,combustible,unidad y fecha
Poniendo el id de su FK para la Base de Datos

### columna categoria

df_categorias

dictio_categoria = crear_dictio(df_categorias,"categoria")
df_historico = df_a_db(dictio_categoria,df_historico,"categoria")
df_historico.sample()

### columna combustible

df_combustibles

dictio_combustibles = crear_dictio(df_combustibles,"combustible")
df_historico = df_a_db(dictio_combustibles,df_historico,"combustible")
df_historico.sample()

### columna unidad

df_unidades

dictio_unidades = crear_dictio(df_unidades,"unidad")
df_historico = df_a_db(dictio_unidades,df_historico,"unidad")
df_historico.sample()

### columna fecha

df_fechas.sample()

dictio_fechas = crear_dictio(df_fechas,"fecha")
df_historico = df_a_db(dictio_fechas,df_historico,"fecha")
df_historico.sample()

# Guardamos y continuamos

df_historico.to_csv("../datos/05_TablasPreDB/05_tabla_historicos.csv")
