==> notebooks/04_base_de_datos.ipynb <==
# Fase DB
En esta fase vamos a crear la base de datos e insertar nuestras tablas.

# Importación de librerías necesarias para trabajar con bases de datos y DataFrames

# Conexión y manejo de bases de datos PostgreSQL
# -----------------------------------------------------------------------
import psycopg2                                   # Permite la conexión y ejecución de consultas en bases de datos PostgreSQL
from psycopg2 import OperationalError, errorcodes, errors  # Maneja errores específicos de psycopg2 para un control preciso de excepciones

# Manipulación de DataFrames
# -----------------------------------------------------------------------
import pandas as pd                               # Manipulación y análisis de datos en estructuras DataFrame, ideal para datos tabulares

# Librerías del sistema y utilidades
# -----------------------------------------------------------------------
import os                                         # Interactúa con el sistema operativo para gestionar rutas y archivos
import sys                                        # Modifica rutas de búsqueda de módulos y gestiona la ejecución del programa
sys.path.append("../")                            # Añade el directorio padre al PATH para permitir la importación de módulos personalizados

# Importación de funciones personalizadas para trabajar con la base de datos
from src.SupportDB import (
    crear_bd,                                     # Crea una base de datos en PostgreSQL si no existe
    conectarse_a_bd,                              # Establece la conexión con la base de datos
    modificar_bd,                                 # Ejecuta comandos SQL para modificar la estructura de la base de datos
    insertar_muchos_datos_bd,                     # Inserta grandes volúmenes de datos en la base de datos
    generar_tupla                                 # Convierte los datos en tuplas para su inserción masiva en la base de datos
)

# Manejo de variables de entorno
# -----------------------------------------------------------------------
import dotenv                                     # Carga variables de entorno desde un archivo .env para la configuración segura de la base de datos
dotenv.load_dotenv()                              # Carga las variables de entorno necesarias para la conexión segura


### Cargamos lo necesario para crear y conectarnos a la base de datos

db = "proyecto5"
user = os.getenv("user")
key = os.getenv("sql")
host = os.getenv("host")
port = os.getenv("port")

### Creamos la base de datos

crear_bd(db,user,key,host,port)

### Nos conectamos a la base de datos

conexion = conectarse_a_bd(db,user,key,host,port)
cursor = conexion.cursor()

# Crear tablas e insertar datos
Ahora crearemos las tablas en el siguiente orden:
- fechas
- combustibles
- categorias
- unidades
- historico_precios

### Tabla fechas
```SQL
    CREATE TABLE fechas (
        id_fecha INT PRIMARY KEY,
        fecha DATE UNIQUE NOT NULL
    );


conexion = conectarse_a_bd(db,user,key,host,port)
query = ''' CREATE TABLE fechas (
            id_fecha INT PRIMARY KEY,
            fecha DATE UNIQUE NOT NULL
        );'''
modificar_bd(conexion,query)

### Insertar fechas

df_fechas = pd.read_csv("../datos/05_TablasPreDB/01_tabla_fechas.csv",index_col="Unnamed: 0")
df_fechas.sample()

conexion = conectarse_a_bd(db,user,key,host,port)
query = ''' INSERT INTO fechas(id_fecha,fecha) VALUES (%s,%s)'''
insertar_muchos_datos_bd(conexion,query,generar_tupla(df_fechas))

### Tabla combustibles
```SQL
    CREATE TABLE combustibles (
        id_combustible INT PRIMARY KEY,
        combustible VARCHAR UNIQUE NOT NULL
    );


conexion = conectarse_a_bd(db,user,key,host,port)
query = ''' CREATE TABLE combustibles (
            id_combustible INT PRIMARY KEY,
            combustible VARCHAR UNIQUE NOT NULL
        );'''
modificar_bd(conexion,query)

### Insertar combustibles

df_combustible = pd.read_csv("../datos/05_TablasPreDB/02_tabla_combustibles.csv",index_col="Unnamed: 0")
df_combustible

df_combustible.fillna("N/A",inplace=True)

conexion = conectarse_a_bd(db,user,key,host,port)
query = ''' INSERT INTO combustibles(id_combustible,combustible) VALUES (%s,%s)'''
insertar_muchos_datos_bd(conexion,query,generar_tupla(df_combustible))

### Tabla categorías
```SQL
    CREATE TABLE categorias (
        id_categoria INT PRIMARY KEY,
        categoria VARCHAR UNIQUE NOT NULL
    );


conexion = conectarse_a_bd(db,user,key,host,port)
query = ''' CREATE TABLE categorias (
            id_categoria INT PRIMARY KEY,
            categoria VARCHAR UNIQUE NOT NULL
        );'''
modificar_bd(conexion,query)

### Insertar categorías

df_categorias = pd.read_csv("../datos/05_TablasPreDB/03_tabla_categorias.csv", index_col="Unnamed: 0")
df_categorias

conexion = conectarse_a_bd(db,user,key,host,port)
query = ''' INSERT INTO categorias(id_categoria,categoria) VALUES (%s,%s)'''
insertar_muchos_datos_bd(conexion,query,generar_tupla(df_categorias))

### Tabla unidades
```SQL
    CREATE TABLE unidades (
        id_unidad INT PRIMARY KEY,
        unidad VARCHAR UNIQUE NOT NULL
    );


conexion = conectarse_a_bd(db,user,key,host,port)
query = ''' CREATE TABLE unidades (
            id_unidad INT PRIMARY KEY,
            unidad VARCHAR UNIQUE NOT NULL
            );'''
modificar_bd(conexion,query)

### Insertar unidades

df_unidades = pd.read_csv("../datos/05_TablasPreDB/04_tabla_unidades.csv",index_col="Unnamed: 0")
df_unidades

conexion = conectarse_a_bd(db,user,key,host,port)
query = ''' INSERT INTO unidades(id_unidad,unidad) VALUES (%s,%s)'''
insertar_muchos_datos_bd(conexion,query,generar_tupla(df_unidades))

### Tabla historico_precios
```SQL
    CREATE TABLE historico_precios (
        id_precio INT PRIMARY KEY,
        id_categoria INT REFERENCES categorias(id_categoria),
        id_combustible INT REFERENCES combustibles(id_combustible),
        precio FLOAT NOT NULL,
        id_unidad INT REFERENCES unidades(id_unidad),
        id_fecha INT REFERENCES fechas(id_fecha)
    );


conexion = conectarse_a_bd(db,user,key,host,port)
query = ''' CREATE TABLE historico_precios (
            id_precio INT PRIMARY KEY,
            id_categoria INT REFERENCES categorias(id_categoria),
            id_combustible INT REFERENCES combustibles(id_combustible),
            precio FLOAT NOT NULL,
            id_unidad INT REFERENCES unidades(id_unidad),
            id_fecha INT REFERENCES fechas(id_fecha)
        );'''
modificar_bd(conexion,query)

### Insertar valores

df_historicos = pd.read_csv("../datos/05_TablasPreDB/05_tabla_historicos.csv", index_col="Unnamed: 0")
df_historicos.sample()

conexion = conectarse_a_bd(db,user,key,host,port)
query = ''' INSERT INTO historico_precios(id_precio,id_categoria,id_combustible,precio,id_unidad,id_fecha)
            VALUES (%s,%s,%s,%s,%s,%s)'''
tupla_historicos = list(df_historicos.itertuples(index=False, name= None))
insertar_muchos_datos_bd(conexion,query,tupla_historicos)
