==> notebooks/05_consultas_pre_visualizacion.ipynb <==
# Importación de librerías necesarias para trabajar con bases de datos y DataFrames

# Conexión y manejo de bases de datos PostgreSQL
# -----------------------------------------------------------------------
import psycopg2                                   # Permite la conexión y ejecución de consultas en bases de datos PostgreSQL
from psycopg2 import OperationalError, errorcodes, errors  # Maneja errores específicos de psycopg2 para un control detallado de excepciones

# Manipulación de DataFrames
# -----------------------------------------------------------------------
import pandas as pd                               # Manipulación y análisis de datos en estructuras DataFrame, ideal para trabajar con datos tabulares

# Librerías del sistema y utilidades
# -----------------------------------------------------------------------
import os                                         # Interactúa con el sistema operativo para gestionar rutas y archivos
import sys                                        # Modifica las rutas de búsqueda de módulos y permite gestionar la ejecución del programa
sys.path.append("../")                            # Añade el directorio padre al PATH para permitir la importación de módulos personalizados

# Importación de funciones personalizadas para trabajar con la base de datos
from src.SupportDB import (
    conectarse_a_bd,                              # Establece la conexión con la base de datos
    consulta_sql                                  # Realiza consultas SQL a la base de datos y devuelve los resultados
)

# Manejo de variables de entorno
# -----------------------------------------------------------------------
import dotenv                                     # Carga variables de entorno desde un archivo .env para configuración segura
dotenv.load_dotenv()                              # Carga las variables de entorno necesarias para la conexión segura

### Cargamos lo necesario para crear y conectarnos a la base de datos

db = "proyecto5"
user = os.getenv("user")
key = os.getenv("sql")
host = os.getenv("host")
port = os.getenv("port")

### Nos conectamos a la base de datos

conexion = conectarse_a_bd(db,user,key,host,port)
cursor = conexion.cursor()

### Evolución Anual del Precio Promedio por Categoría

conexion = conectarse_a_bd(db,user,key,host,port)
query = """ SELECT 
                categoria,
                ROUND(AVG(precio)::NUMERIC,2) AS precio_promedio,
                EXTRACT(YEAR FROM fecha) AS anio
            FROM historico_precios
            NATURAL JOIN categorias
            NATURAL JOIN fechas
            GROUP BY categoria, anio
            ORDER BY categoria, anio;"""

df_consulta1 = consulta_sql(conexion,query)
df_consulta1.to_csv("../datos/06_TablasPreVisualizacion/01_evolucion_anual_precio_promedio_por_categoria.csv")
df_consulta1.sample()

### Variación Mensual de Precio por categoría

conexion = conectarse_a_bd(db,user,key,host,port)
query = """ SELECT 
                categoria,
                ROUND(AVG(precio)::NUMERIC,2) AS precio_promedio,
                fecha
            FROM historico_precios
            NATURAL JOIN categorias
            NATURAL JOIN fechas
            GROUP BY categoria,fecha
            ORDER BY categoria,fecha;"""

df_consulta2 = consulta_sql(conexion,query)
df_consulta2.to_csv("../datos/06_TablasPreVisualizacion/02_variacion_mensual_precio_por_categoria.csv")
df_consulta2.sample()

### Variacion Mensual de la Luz

conexion = conectarse_a_bd(db,user,key,host,port)
query = """ SELECT 
                categoria,
                fecha,
            ROUND(AVG(precio)::NUMERIC, 2) AS precio_promedio
            FROM historico_precios
            NATURAL JOIN categorias
            NATURAL JOIN fechas
            WHERE categoria = 'luz'
            GROUP BY categoria, fecha
            ORDER BY fecha;"""
df_consulta3 = consulta_sql(conexion,query)
df_consulta3.to_csv("../datos/06_TablasPreVisualizacion/03_variacion_mensual_luz.csv")
df_consulta3.sample()

### Variacion Mensual del gas

conexion = conectarse_a_bd(db,user,key,host,port)
query = """ SELECT 
                categoria,
                fecha,
                ROUND(AVG(precio)::NUMERIC, 2) AS precio_promedio
            FROM historico_precios
            NATURAL JOIN categorias
            NATURAL JOIN fechas
            WHERE categoria = 'gas'
            GROUP BY categoria, fecha
            ORDER BY fecha;"""
df_consulta4 = consulta_sql(conexion,query)
df_consulta4.to_csv("../datos/06_TablasPreVisualizacion/04_variacion_mensual_gas.csv")
df_consulta4.sample()

### Variación Mensual de Precio por Combustible

conexion = conectarse_a_bd(db,user,key,host,port)
query = """ SELECT 
                combustible,
                AVG(precio) AS precio_promedio,
                fecha
            FROM historico_precios
	        NATURAL JOIN combustibles
	        NATURAL JOIN fechas
	        NATURAL JOIN categorias
            WHERE categoria = 'combustible'
            GROUP BY combustible, fecha
            ORDER BY combustible, fecha;"""
df_consulta5 = consulta_sql(conexion,query)
df_consulta5.to_csv("../datos/06_TablasPreVisualizacion/05_variacion_mensual_precio_combustible.csv")
df_consulta5.sample()

### Variacion Mensual del Agua

conexion = conectarse_a_bd(db,user,key,host,port)
query = """ SELECT 
                categoria,
                fecha,
                ROUND(AVG(precio)::NUMERIC, 2) AS precio_promedio
            FROM historico_precios
            NATURAL JOIN categorias
            NATURAL JOIN fechas
            WHERE categoria = 'agua'
            GROUP BY categoria, fecha
            ORDER BY fecha;"""
df_consulta6 = consulta_sql(conexion,query)
df_consulta6.to_csv("../datos/06_TablasPreVisualizacion/06_variacion_mensual_agua.csv")
df_consulta6.sample()
